requirements.txt
paramiko
schedule>=0.6.0
paho-mqtt>=1.5.0
